requirements.txt
numpy==1.14.2
PyQt5==5.10.1
pyqtgraph==0.10.0
pyserial==3.4
sip==4.19.8
